Fix #3421: Separate MarkStoriesCompletedActivityTest into activity and fragment test files
